# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "CompManualAdd": {
        "description": "POST /acquisitions/{id}/comps \u2014 manual add by URL or direct fields.",
        "properties": {
          "avg_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Rate"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "CompManualAdd",
        "type": "object"
      },
      "CompOut": {
        "properties": {
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "amenity_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenity Rank"
          },
          "amenity_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenity Score"
          },
          "avg_rate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Rate"
          },
          "best_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Snippet"
          },
          "comp_id": {
            "title": "Comp Id",
            "type": "string"
          },
          "distance_mi": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Mi"
          },
          "is_manual": {
            "default": false,
            "title": "Is Manual",
            "type": "boolean"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "worst_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worst Snippet"
          }
        },
        "required": [
          "comp_id",
          "name"
        ],
        "title": "CompOut",
        "type": "object"
      },
      "CompScatterPoint": {
        "properties": {
          "amenity_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenity Score"
          },
          "avg_rate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Rate"
          },
          "comp_id": {
            "title": "Comp Id",
            "type": "string"
          },
          "is_target": {
            "default": false,
            "title": "Is Target",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          }
        },
        "required": [
          "comp_id",
          "name"
        ],
        "title": "CompScatterPoint",
        "type": "object"
      },
      "CompSet": {
        "properties": {
          "comps": {
            "items": {
              "$ref": "#/components/schemas/CompOut"
            },
            "title": "Comps",
            "type": "array"
          },
          "visualization": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompVisualization"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "CompSet",
        "type": "object"
      },
      "CompVisualization": {
        "description": "Rate\u00d7sentiment / rate\u00d7amenities scatter + ranked list for the Comps tab.",
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/CompScatterPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "title": "CompVisualization",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Structured error envelope (CLAUDE.md API design).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API surface per design doc \u00a79. Phase-0 endpoints are typed contract stubs; domain logic lands in Phases 1\u20134.",
    "title": "RJourney Acquisitions Platform API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/acquisitions/{acquisition_id}/comps": {
      "get": {
        "description": "Comp set + visualization data (rate\u00d7sentiment / rate\u00d7amenities).",
        "operationId": "get_comps_acquisitions__acquisition_id__comps_get",
        "parameters": [
          {
            "in": "path",
            "name": "acquisition_id",
            "required": true,
            "schema": {
              "title": "Acquisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Get Comps",
        "tags": [
          "comps"
        ]
      },
      "post": {
        "description": "Manual competitor add by URL or direct fields \u2192 scrape/enrich.",
        "operationId": "add_comp_acquisitions__acquisition_id__comps_post",
        "parameters": [
          {
            "in": "path",
            "name": "acquisition_id",
            "required": true,
            "schema": {
              "title": "Acquisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompManualAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Add Comp",
        "tags": [
          "comps"
        ]
      }
    },
...
  }
}